requirements.txt
django
djangorestframework
psycopg2
drf-spectacular
channels
channels-redis
daphne